Moving the bow api into the item api
