--- fabric-combat-api-v1/build.gradle
@@ -1,10 +1,0 @@
-archivesBaseName = "fabric-combat-api-v1"
-version = getSubprojectVersion(project, "1.0.0")
-
-dependencies {
-	testmodCompile project(path: ':fabric-object-builder-api-v1', configuration: 'dev')
-}
-
-moduleDependencies(project, [
-		'fabric-api-base'
-])
