--- fabric-combat-api-v1/src/main/resources/fabric-combat-api-v1.mixins.json
@@ -1,20 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.combat",
-  "compatibilityLevel": "JAVA_8",
-  "mixins": [
-    "AbstractSkeletonEntityMixin",
-    "BowAttackGoalMixin",
-    "BowAttackMixin",
-    "BowItemMixin",
-    "EntityUseBowMixin",
-    "ProjectileUtilMixin"
-  ],
-  "client": [
-    "client.HeldItemRendererMixin",
-    "client.SkeletonEntityModelMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
