--- fabric-combat-api-v1/src/main/resources/fabric.mod.json
@@ -1,29 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-combat-api-v1",
-  "name": "Fabric Combat API (v1)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-combat-api-v1/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "mixins": [
-    "fabric-combat-api-v1.mixins.json"
-  ],
-  "depends": {
-    "fabricloader": ">=0.4.0",
-    "fabric-api-base": "*"
-  },
-  "description": "Hooks for combat related events and items",
-  "custom": {
-    "fabric-api:module-lifecycle": "stable"
-  }
-}
