--- fabric-combat-api-v1/src/testmod/resources/fabric.mod.json
@@ -1,19 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-combat-api-v1-testmod",
-  "name": "Fabric Combat API (v1) Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-item-api-v1": "*"
-  },
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.test.combat.FabricBowTests"
-    ],
-    "client": [
-      "net.fabricmc.fabric.test.combat.client.FabricBowTestmodClient"
-    ]
-  }
-}
