+++ fabric-item-api-v1/src/testmod/resources/fabric.mod.json
@@ -10,13 +10,13 @@
   },
   "entrypoints": {
     "main": [
-      "net.fabricmc.fabric.test.item.FabricItemSettingsTests"
+      "net.fabricmc.fabric.test.item.FabricItemSettingsTests",
+      "net.fabricmc.fabric.test.item.CustomDamageTest",
+      "net.fabricmc.fabric.test.item.FabricBowTests"
     ],
     "client": [
-      "net.fabricmc.fabric.test.item.client.TooltipTests"
-    ],
-    "main": [
-      "net.fabricmc.fabric.test.item.CustomDamageTest"
+      "net.fabricmc.fabric.test.item.client.TooltipTests",
+      "net.fabricmc.fabric.test.item.client.FabricBowTestClient"
     ]
   }
 }
